block details done

+++ app/src/main/java/com/tyshko/getblock/models/stack/Block.kt
@@ -1,10 +1,10 @@
 package com.tyshko.getblock.models.stack
 
-data class Block (
+data class Block(
     val block: Long,
     val signature: String,
     val time: Long,
     val epoch: Int,
-    val rewardLamports: Long,
+    val rewardLamports: Int,
     val previousBlockHash: String
 )
